+++ Application/Models/Тreatments/AddТreatmentModel.cs
@@ -20,7 +20,7 @@ namespace Application.Models.Тreatments
 
         public int? AmountOfFuel { get; init; }
 
-        public int? FuelPrice { get; init; }
+        public decimal? FuelPrice { get; init; }
 
         [Required]
         public int ArticleId { get; init; }
@@ -29,6 +29,6 @@ namespace Application.Models.Тreatments
         public int ArticleQuantity { get; init; }
 
         [Required]
-        public int ArticlePrice { get; init; }
+        public decimal ArticlePrice { get; init; }
     }
 }
